fix(databass): Stop using `UserReport` foreign key relations, and replace with `_id` instead.
We're moving `UserReport` to a separate database, and so can no longer rely on  Django's
automatic join behaviour. In https://github.com/getsentry/sentry/pull/23194 we will remove teh
foreign keys entirely and replace with just the `_id` columns instead. This pr prepares us for that,
and attempts to remove all join behaviour as well as replacing uses of the FKs with `*_id`.

Not 100% done - we have a cross db join to groups in a couple places. I'll figure out how to handle
that shortly.